+++ src/ui/editor/components/DropdownInput.tsx
@@ -1,4 +1,5 @@
 import {Autocomplete, type SxProps, TextField, type Theme} from "@mui/material";
+import * as React from "react";
 
 type DropdownInputProps = {
   disabled?: boolean
