+++ src/ui/editor/editors/ItemStackEditor/index.ts
@@ -1,2 +1,1 @@
 export {default as ItemStackEditor} from './ItemStackEditor';
-export type { ItemStackEditorResult } from './ItemStackEditor';
